% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{draw.struct}
\alias{draw.struct}
\title{Draw 3D structure}
\usage{
draw.struct(coordinates)
}
\arguments{
\item{coordinates}{A matrix of three columns corresponding to x, y, and z coordinates.}
}
\value{
The figure.
}
\description{
This function provides visualization of the 3D structure through the rgl package.
}
