% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{get.breakpoints}
\alias{get.breakpoints}
\title{Generate overlapping block boundaries}
\usage{
get.breakpoints(breaks = NULL, n, block_size = 40, noverlap = 1)
}
\arguments{
\item{breaks}{A vector that describes the indices of pre-specified cut points. If unspecified, equal block size cuts will be used.}

\item{n}{The number of loci to partition.}

\item{block_size}{The size of each block.}

\item{noverlap}{Number of loci that overlap between neighboring blocks.}
}
\value{
A list. Each element is a size-2 vector detailing the index of the first and last element of the block.
}
\description{
Generate overlapping block boundaries
}
