% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{get.sample}
\alias{get.sample}
\title{Get posterior samples of coordinates for a specific locus}
\usage{
get.sample(psample, i)
}
\arguments{
\item{psample}{An output from tRex().}

\item{i}{The index for which to retrive coordinates.}
}
\value{
A matrix of 10,000 rows. 1st, 2nd, and 3rd columns corresponds to x,y, and z coordinates. Each row corresponds to a posterior sample  of coordinates.
}
\description{
Get posterior samples of coordinates for a specific locus
}
