% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sim_bias}
\alias{sim_bias}
\title{Simulated bias matrix}
\format{
A matrix with n rows and 3 columns:
\describe{
\item{column 1}{effective fragment information.}
\item{column 2}{GC content}
\item{column 3}{mappability}
}
}
\usage{
sim_bias
}
\description{
Simulated bias matrix
}
\keyword{datasets}
