\name{sim_bias_human}
\alias{sim_bias_human}
\docType{data}
\title{
	Simulated bias data for human1422.
}
\description{
}
\usage{data("sim_bias_human")}
\format{
The sim_bias is data of matrix format with following column names:
Effective.Length, GC.contents, Mappability.
}
\examples{
data(sim_bias_human)
}
\keyword{datasets}
