% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sim_hic}
\alias{sim_hic}
\title{Simulated Hi-C contact matrix}
\format{
A \eqn{n \times n} matrix where the (i, j) element is the interaction count between bin i and bin j.
The matrix is symmetric with 0 on the diagonal.
}
\usage{
sim_hic
}
\description{
Simulated Hi-C contact matrix
}
\keyword{datasets}
