% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{simulate_contact}
\alias{simulate_contact}
\title{Simulate Count Matrix based on Known Structure}
\usage{
simulate_contact(S, family = "poisson", pi = 0, p = 0.75, random = FALSE)
}
\arguments{
\item{S}{The coordinates: an \eqn{n \times 3} matrix, where each row represent the (x, y, z) coordinates of one locus.}

\item{family}{The distribution for simulation. One of "poisson" or "NB". Default to "poisson".}

\item{pi}{Zero inflation parameter for zeo inflated models.  Default is 0 (no zero inflation).}

\item{p}{Overdispersion parameter for negative binomial. Deafult is 0.75.}

\item{random}{Whether to add random effect. Default is FALSE.}
}
\value{
A symmetric \eqn{n \times n} count matrix.
}
\description{
Simulate Count Matrix based on Known Structure
}
