% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{Cut}
\alias{Cut}
\title{Run the cut part of cut and paste}
\usage{
Cut(
  contact,
  bias = NULL,
  breaks = NULL,
  block_size = 40,
  noverlap = 1,
  CPU = 1,
  save_mcmc = FALSE
)
}
\arguments{
\item{contact}{placeholder}

\item{bias}{placeholder}

\item{breaks}{placeholder}

\item{block_size}{placeholder}

\item{noverlap}{placeholder}

\item{CPU}{Integer specifying the number of cores for parallel MCMC execution. Default to 1.}

\item{save_mcmc}{Whether to save the MCMC result. Default is FALSE.}
}
\value{
A list...
}
\description{
Run the cut part of cut and paste
}
