% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{Cut}
\alias{Cut}
\title{Run Cut Algorithm}
\usage{
Cut(
  contact,
  bias = NULL,
  breaks = NULL,
  block_size = 40,
  noverlap = 1,
  CPU = 1,
  save_mcmc = FALSE
)
}
\arguments{
\item{contact}{The contact matrix: an \eqn{n \times n} matrix, where n is the number of loci. Its element (i, j) denotes the number of interactions between locus i and j.}

\item{bias}{placeholder}

\item{breaks}{A vector that describes the index of pre-specified cut points. If unspecified, equal block size cuts will be used.}

\item{block_size}{The size of each block.}

\item{noverlap}{Number of loci that overlaps between neighboring blocks.}

\item{CPU}{Integer specifying the number of cores for parallel MCMC execution. Default to 1.}

\item{save_mcmc}{Whether to save the posterior samples generated during the MCMC step. Default to FALSE.}
}
\value{
A list of the following elements: cutlist, result, noverlap and block_size.
}
\description{
Cut the contact matrix into parts and estimate the structure of each partition.
}
