% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{CutAndPaste}
\alias{CutAndPaste}
\title{Run Cut-And-Paste}
\usage{
CutAndPaste(
  contact,
  bias = NULL,
  breaks = NULL,
  block_size = 40,
  noverlap = 1,
  CPU,
  save_mcmc = FALSE,
  mcmc = 1e+05,
  burn = 1e+05,
  thinning = 10,
  jump.beta1 = 0.05,
  jump.x = 0.2,
  jump.u = 0.2,
  jump.cov1 = 0.1,
  jump.cov2 = 0.1,
  leapfrog.L = 10,
  leapfrog.e = 0.001,
  target.accept = 0.95,
  tuning_control = NULL,
  method.type = "tRex"
)
}
\arguments{
\item{contact}{The contact matrix: an \eqn{n \times n} matrix, where n is the number of loci. Its element (i, j) denotes the number of interactions between locus i and j.}

\item{bias}{HiC bias matrix. An \eqn{n \times 3} matrix whose columns are: effective fragment information, GC content, and mappability.}

\item{breaks}{A vector that describes the indices of pre-specified cut points. If unspecified, equal block size cuts will be used.}

\item{block_size}{The size of each block.}

\item{CPU}{Integer specifying the number of cores for parallel MCMC execution. Defaults to 1.}

\item{save_mcmc}{Whether to save posterior samples generated during Cut(). Defaults to FALSE.}
}
\value{
An \eqn{n \times 3} matrix of the estimated coordinates.
}
\description{
Run Cut-And-Paste
}
