% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{CutAndPaste}
\alias{CutAndPaste}
\title{Run both cut and paste}
\usage{
CutAndPaste(
  contact,
  bias = NULL,
  breaks = NULL,
  block_size = 40,
  noverlap = 1,
  CPU,
  save_mcmc = FALSE
)
}
\arguments{
\item{contact}{placeholder}

\item{CPU}{placeholder}

\item{save_mcmc}{placeholder}

\item{cutresult}{placeholder}
}
\value{
A list...
}
\description{
Run both cut and paste
}
