% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{CutAndPaste}
\alias{CutAndPaste}
\title{Run both cut and paste in one step}
\usage{
CutAndPaste(
  contact,
  bias = NULL,
  breaks = NULL,
  block_size = 40,
  noverlap = 1,
  CPU,
  save_mcmc = FALSE
)
}
\arguments{
\item{contact}{The contact matrix: an \eqn{n \times n} matrix, where n is the number of loci. Its element (i, j) denotes the number of interactions between locus i and j.}

\item{bias}{placeholder}

\item{CPU}{Integer specifying the number of cores for parallel MCMC execution. Default to 1.}

\item{save_mcmc}{Whether to save posterior samples generated during Cut(). Default to FALSE.}
}
\value{
A list...
}
\description{
Run both cut and paste in one step
}
