% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{draw.struct}
\alias{draw.struct}
\title{Draw 3D structure}
\usage{
draw.struct(coordinates)
}
\arguments{
\item{coordinates}{A matrix of three coloumns correspoding to three coordinates.}
}
\value{
The figure.
}
\description{
Using a rgl package, this function provides a visualization function of the 3D structure
}
