% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{get.param}
\alias{get.param}
\title{get posterior samples of parameters.}
\usage{
get.param(psample)
}
\arguments{
\item{psample}{An output of tRex().}
}
\value{
A matrix of 10000 rows and 5 columns of \beta_1, cov_1, cov_2, \sigma_x, and \sigma_u.
}
\description{
get posterior samples of parameters.
}
