% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{get.sample}
\alias{get.sample}
\title{get posterior samples of coordinates of a specific locus.}
\usage{
get.sample(psample, i)
}
\arguments{
\item{psample}{An output of tRex().}

\item{i}{The locus to get coordinates.}
}
\value{
A Matrix of 10000 rows. 1st, 2nd, and 3rd colums corresponds to x,y, and z coordinates. Each rows corresponds to a posterior sample  of coordinates.
}
\description{
get posterior samples of coordinates of a specific locus.
}
