% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{Paste}
\alias{Paste}
\title{Run the paste part of cut and paste}
\usage{
Paste(contact, cutresult, CPU)
}
\arguments{
\item{contact}{The contact matrix: an \eqn{n \times n} matrix, where n is the number of loci. Its element (i, j) denotes the number of interactions between locus i and j.}

\item{cutresult}{The result from a Cut() call.}

\item{CPU}{Integer specifying the number of cores for parallel MCMC execution. Defaults to 1.}
}
\value{
An \eqn{n \times 3} matrix of the estimated coordinates.
}
\description{
Run the paste part of cut and paste
}
