% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{Paste}
\alias{Paste}
\title{Run the paste part of cut and paste}
\usage{
Paste(contact, cutresult, CPU = 1)
}
\arguments{
\item{contact}{The contact matrix. See Cut() for detailed explanation.}

\item{cutresult}{The result object from running Cut().}

\item{CPU}{Integer specifying the number of cores for parallel MCMC execution. Default to 1.}
}
\value{
A list of estimated 3D coordinates and likelihood.
}
\description{
Run the paste part of cut and paste
}
