% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRex.r
\name{Paste}
\alias{Paste}
\title{Run the paste part of cut and paste}
\usage{
Paste(contact, cutresult, block_size, CPU = 1)
}
\arguments{
\item{contact}{placeholder}

\item{cutresult}{placeholder}

\item{CPU}{Integer specifying the number of cores for parallel MCMC execution. Default to 1.}
}
\value{
A list...
}
\description{
Run the paste part of cut and paste
}
