\name{sim_bias}
\alias{sim_bias}
\docType{data}
\title{
	Simulated bias data for sim_hic.
}
\description{
}
\usage{data("sim_bias")}
\format{
The sim_bias is data of matrix format with following column names:
Effective.Length, GC.contents, Mappability.
}
\examples{
data(sim_bias)
}
\keyword{datasets}
